requirements.txt
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
Django==1.7
djangorestframework==2.4.8
drf-httpsig==1.2.0
httpsig==1.3.0
idna==3.4
paramiko==3.3.1
pycparser==2.21
pycryptodome==3.18.0
PyNaCl==1.5.0
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
tqdm==4.66.1
urllib3==2.0.4
